depth/lift: 64-bit single-axis universal bigWig (retire 2e9 chunking)
Phase 1c: move the universal-depth bigWig from the 32-bit wigToBigWig + 2e9
uni0..uniN chunking workaround to a single 64-bit uni0 axis via the vendored
64-bit libBigWig fork.

- vendored libBigWig: the 7 fork files (64-bit coords + zoom + reader gating).
- taf_lift.c: --bigwig reads the single 64-bit uni0 axis (base=0, u64 query);
  deleted the chunk pre-split + chunk-offset mapping; the axis guard now
  requires a single uni0 chrom (nKeys==1) and rejects pre-1c multi-chrom files.
- taf_depth.c: NEW --bigwig <file> writes the binned depth directly as a 64-bit
  bigWig (single uni0/T axis) via the fork's bwAdd*, replacing external
  wigToBigWig (which can't represent a >2^31 axis). Hooked into the sorted
  single-threaded binner flush.
- tui.c/tui.h: removed tui_split_intervals_on_chunks + TUI_UNI_CHUNK (a pure
  32-bit workaround; the .tui format never encoded the 2e9 axis).
- depth_test.c: dropped the split-on-chunks test; CLI test asserts single uni0.

Validated: depth bedGraph bit-identical to the old chunked build over the whole
apes genome (crosses the old 2e9 boundary); depth --bigwig -> lift --bigwig
round-trip correct (non-zero; --depth vs --bigwig byte-identical); 32-bit and
multi-chrom .bw rejected cleanly. 97 CuTest + 1021 pytest pass.

Follow-ups (not here): the SLURM/local depth scripts still drive external
wigToBigWig and need updating to --bigwig (+ a 64-bit per-shard merge for
--columnRange sharding); a few stale wigToBigWig comments + the design doc.

+++ taffy/submodules/libBigWig/bwCommon.h
@@ -57,7 +57,7 @@ void bwDestroyIndexNode(bwRTreeNode_t *node);
 void bwDestroyIndex(bwRTree_t *idx);
 
 /// @cond SKIP
-bwOverlapBlock_t *walkRTreeNodes(bigWigFile_t *bw, bwRTreeNode_t *root, uint32_t tid, uint32_t start, uint32_t end);
+bwOverlapBlock_t *walkRTreeNodes(bigWigFile_t *bw, bwRTreeNode_t *root, uint32_t tid, uint64_t start, uint64_t end);
 void destroyBWOverlapBlock(bwOverlapBlock_t *b);
 /// @endcond
